+++ changelog.md
@@ -12,6 +12,7 @@ If you are reading this on the website, then the features listed are live here.
 - [`spawn`](https://uiua.org/docs/spawn) and [`wait`](https://uiua.org/docs/wait) no longer have glyphs. Code using `↰` and `↲` will continue to work and will be formatted as `spawn` and `wait`.
 - `&n` is not longer a system function and is now called [`now`](https://uiua.org/docs/now)
 - [`under` `⍜`](https://uiua.org/docs/under) [`now`](https://uiua.org/docs/now) can be used to time things
+- [`call` `!`](https://uiua.org/docs/call) can now call functions that return any number of values, not just one
 ### Interpreter
 - The formatter now aligns consecutive end-of-line comments
 - Fix a bug that prevented [`under` `⍜`](https://uiua.org/docs/under) multidimensional [`take` `↙`](https://uiua.org/docs/take) and [`drop` `↘`](https://uiua.org/docs/drop) from working